Deprecating old scripts with new scripts following successful test run of refactored code
